requirements.txt
astroid==2.4.2
branca==0.4.2
certifi==2021.10.8
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
configparser==5.0.2
Flask==2.0.1
folium==0.12.1
future==0.18.2
graphviz==0.17
idna==3.2
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
mccabe==0.6.1
numpy==1.21.2
pandas==1.3.3
polyline==1.4.0
pylint==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
requests==2.26.0
six==1.16.0
toml==0.10.2
transitions==0.8.10
urllib3==1.26.6
Werkzeug==2.0.1
wincertstore==0.2
wrapt==1.12.1
gunicorn==19.4.5